+++ packages/websocket/test/controller/io_test.dart
@@ -9,7 +9,7 @@ import 'package:test/test.dart';
 import 'common.dart';
 
 void main() {
-  srv.Angel? app;
+  srv.Angel app;
   late srv.AngelHttp http;
   ws.WebSockets? client;
   srv.AngelWebSocket websockets;
@@ -18,17 +18,17 @@ void main() {
 
   setUp(() async {
     app = srv.Angel(reflector: const MirrorsReflector());
-    http = srv.AngelHttp(app!, useZone: false);
+    http = srv.AngelHttp(app, useZone: false);
 
     websockets = srv.AngelWebSocket(app)
       ..onData.listen((data) {
         print('Received by server: $data');
       });
 
-    await app!.configure(websockets.configureServer);
-    app!.all('/ws', websockets.handleRequest);
-    await app!.configure(GameController(websockets).configureServer);
-    app!.logger = Logger('angel_auth')..onRecord.listen(print);
+    await app.configure(websockets.configureServer);
+    app.all('/ws', websockets.handleRequest);
+    await app.configure(GameController(websockets).configureServer);
+    app.logger = Logger('angel_auth')..onRecord.listen(print);
 
     server = await http.startServer();
     url = 'ws://${server!.address.address}:${server!.port}/ws';
@@ -53,7 +53,7 @@ void main() {
   tearDown(() async {
     await client!.close();
     await http.close();
-    app = null;
+    //app = null;
     client = null;
     server = null;
     url = null;